kopfmysql/requirements.txt
kopf
kubernetes
mysql-connector-python
